feat: update modules: add k8s application update
Add a recipe for the Mender Application Update Module,
specifically the k3s/k8s form. It does not have a direct
dependency on kubernetes respectively a form of it, because
meta-virtualization has various means of providing it and
those do not immediately swap in and out.

Changelog: Title
Ticket: None

Signed-off-by: Josef Holzmayr <jester@theyoctojester.info>